requirements.txt
kivy~=2.1.0
pyttsx3~=2.90
vosk~=0.3.32
sounddevice~=0.4.4
git+https://github.com/Cupcakus/pafy
python-vlc
youtube-dl